+++ src/Search/SearchWithSuggestions/SuggestionItem.module.scss
@@ -1,0 +1,20 @@
+@import '../../mixins';
+
+.suggestion {
+  padding: 6px 8px;
+  color: var(--mirage);
+  height: auto;
+  min-height: 32px;
+  text-align: left;
+
+  @include text('body-3');
+
+  &:hover,
+  &:focus {
+    color: var(--jungle-green);
+  }
+}
+
+.cursored {
+  background: var(--athens);
+}
